Separate task list logic from B33PBOP.java
As the functionality of task list scale, it would be better to separate the concerns of task list actions from that of the bot actions.

I created a new TaskList.java file to handle the task functions while keeping the response functionality within the bot itself.

Along with these changes, I updated the necessary methods and exception handling to fit this new structure.

TODO:
* Update the comments of new methods I created and old methods I changed/moved