Code Refactor: Improve code readability
Previously, the execute method of command classes did not have a prominent happy path. Instead, it made use of if, else blocks.

Guard clauses were moved to the front which allowed me to remove the else block, reducing indentation and improving readability of the code.

The existsInTaskDescription method of EventTask.java had a complicated return statement with nested function calls and 2 or operators.

This made the conditions difficult to follow and understand.

Each nested function call was broken down and stored into several intermediate variables.

This helped simplify the return statement into just 2 operands and a or operator which improves readability.